pyproject.toml
[tool.poetry]
name = "brewblox-ispindel"
version = "0.4.0"
description = "BrewBlox Service for the iSpindel Hydrometer"
authors = ["Benoit Delbosc <bdelbosc@free.com>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.7"
brewblox-service = "^0.28.0"

[tool.poetry.dev-dependencies]
pytest-flake8 = "^1.0.4"
pytest-cov = "^2.8.1"
pytest-mock = "^2.0.0"
pytest-aiohttp = "^0.3.0"
flake8 = "^3.7.9"
autopep8 = "^1.5"
flake8-quotes = "^3.0.0"
aresponses = "^2.0.0"
asyncmock = "^0.4.2"
pytest = "^5.4.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
